// srm_ui/src/app/main/Models/etendering/ViewModels/rfq-supplier-header-country-origin-model.ts
export class RFQSupplierHeaderCountryOriginModel {
    id: string;
    rfqSupplierHeaderInformationId: string;
    countryId: string;
    cityId: string;
    portId: string;
    ifsIncoTermId: string;
    incoTermCountryId: string;
    incoTermCityId: string;
    incoTermPortId: string;
    countryName: string;
    cityName: string;
    portName: string;
    ifsIncoTermName: string;
    isDeleting: boolean;
}